+++ lib/streamlit/time_util.py
@@ -18,7 +18,7 @@ import math
 from datetime import date, timedelta
 from typing import Literal, overload
 
-from streamlit.errors import MarkdownFormattedException, StreamlitAPIException
+from streamlit.errors import BadTimeStringError, StreamlitAPIException
 
 
 def adjust_years(input_date: date, years: int) -> date:
@@ -36,17 +36,6 @@ def adjust_years(input_date: date, years: int) -> date:
             f"Date {input_date} does not exist in the target year {input_date.year + years}. "
             "This should never happen. Please report this bug."
         ) from err
-
-
-class BadTimeStringError(StreamlitAPIException):
-    """Raised when a bad time string argument is passed."""
-
-    def __init__(self, t: str):
-        MarkdownFormattedException.__init__(
-            self,
-            "Time string doesn't look right. It should be formatted as"
-            f"`'1d2h34m'` or `2 days`, for example. Got: {t}",
-        )
 
 
 @overload
